Fix AY-3-8912 prescaler and add --disk-a/b CLI flags
The AY chip has an internal ÷8 prescaler before its tone, noise, and
envelope counters. Multiply all counter thresholds by 8 so register
values match real CPC hardware (e.g. SOUND 1,142,... plays A4 440 Hz
instead of 3 octaves too high). env_counter widened to u32 to handle
the larger values.

Add --disk-a=PATH and --disk-b=PATH CLI flags to override drive images
without editing the config file (useful for testing DSK files).

+++ src/psg.h
@@ -25,7 +25,7 @@ typedef struct {
 
     /* Envelope */
     u16 env_period;
-    u16 env_counter;
+    u32 env_counter;
     u8  env_step;       /* 0-31: 32 steps per cycle */
     bool env_hold;
     bool env_dir;
